# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treasury/transactions:
    get:
      operationId: listTreasuryTransactions
      tags: [treasury]
      summary: Treasury transaction ledger
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: treasuryAccountId
          in: query
          schema: { type: string, format: uuid }
        - name: direction
          in: query
          schema: { type: string }
        - name: referenceType
          in: query
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string }
        - name: dateTo
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasuryTransactionList"
  /treasury/sessions:
    get:
      operationId: listTreasurySessions
      tags: [treasury]
      summary: Treasury session history
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: treasuryAccountId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasurySessionList"
    post:
      operationId: openTreasurySession
      tags: [treasury]
      summary: Open a treasury session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenSessionInput"
      responses:
        "201":
          description: Opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasurySession"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /treasury/sessions/current:
    get:
      operationId: getCurrentTreasurySession
      tags: [treasury]
      summary: Current open session for an account
      parameters:
        - name: treasuryAccountId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current session or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentSessionResponse"
  /treasury/sessions/{id}/close:
    post:
      operationId: closeTreasurySession
      tags: [treasury]
      summary: Close a treasury session
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseSessionInput"
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreasurySession"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TreasuryTransaction:
      type: object
      required: [id, treasuryAccountId, direction, amount, balanceAfter, referenceType, createdAt]
      properties:
        id: { type: string, format: uuid }
        treasuryAccountId: { type: string, format: uuid }
        accountName: { type: ["string", "null"] }
        sessionId: { type: ["string", "null"] }
        direction: { type: string }
        amount: { type: string }
        balanceAfter: { type: string }
        referenceType: { type: string }
        referenceId: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    TreasuryTransactionList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TreasuryTransaction" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    TreasurySession:
      type: object
      required: [id, treasuryAccountId, status, openingBalance, openedAt]
      properties:
        id: { type: string, format: uuid }
        treasuryAccountId: { type: string, format: uuid }
        accountName: { type: ["string", "null"] }
        status: { type: string }
        openingBalance: { type: string }
        expectedClosingBalance: { type: ["string", "null"] }
        actualClosingBalance: { type: ["string", "null"] }
        variance: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        openedByName: { type: ["string", "null"] }
        closedByName: { type: ["string", "null"] }
        openedAt: { type: string, format: date-time }
        closedAt: { type: ["string", "null"] }
    TreasurySessionList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TreasurySession" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    CurrentSessionResponse:
      type: object
      required: [session]
      properties:
        session:
          oneOf:
            - $ref: "#/components/schemas/TreasurySession"
            - type: "null"
    OpenSessionInput:
      type: object
      required: [treasuryAccountId, openingBalance]
      properties:
        treasuryAccountId: { type: string, format: uuid }
        openingBalance: { type: number, minimum: 0 }
        notes: { type: ["string", "null"] }
    CloseSessionInput:
      type: object
      required: [actualClosingBalance]
      properties:
        actualClosingBalance: { type: number, minimum: 0 }
        notes: { type: ["string", "null"] }
...